Cleanup pkg/schema
