+++ pkg/ocw/parse_test.go
@@ -1,6 +1,7 @@
 package ocw
 
 import (
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -89,4 +90,27 @@ func TestParseString_InvalidYAML(t *testing.T) {
 	ocw, err := ParseString(content)
 	require.Error(t, err)
 	assert.Nil(t, ocw)
+}
+
+func TestValidateAndParseFile(t *testing.T) {
+	// Create a temp file for testing
+	content := `
+schemaVersion: v1
+name: test
+jobs:
+  test:
+    sequence:
+      - image: alpine
+        run: echo test
+`
+	tmpFile := t.TempDir() + "/test.yaml"
+	err := os.WriteFile(tmpFile, []byte(content), 0644)
+	assert.NoError(t, err)
+
+	_, err = ValidateAndParseFile(tmpFile)
+	assert.NoError(t, err)
+
+	// Test non-existent file
+	_, err = ValidateAndParseFile("/non/existent/file.yaml")
+	assert.Error(t, err)
 }
