+++ pkg/schema/steps.go
@@ -40,115 +40,8 @@ type OptionalStepBase struct {
 	Needs []string `yaml:"needs,omitempty" json:"needs,omitempty"`
 }
 
-// StringOrStringSlice can be either a single string or a slice of strings
-type StringOrStringSlice struct {
-	Single   *string
-	Multiple []string
-}
-
-// UnmarshalYAML implements custom unmarshaling for StringOrStringSlice
-func (s *StringOrStringSlice) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var single string
-	if err := unmarshal(&single); err == nil {
-		s.Single = &single
-		return nil
-	}
-
-	var multiple []string
-	if err := unmarshal(&multiple); err == nil {
-		s.Multiple = multiple
-		return nil
-	}
-
-	return nil
-}
-
-// MarshalYAML implements custom marshaling for StringOrStringSlice
-func (s StringOrStringSlice) MarshalYAML() (interface{}, error) {
-	if s.Single != nil {
-		return *s.Single, nil
-	}
-	return s.Multiple, nil
-}
-
-// StringMapOrSlice can be either a map[string]string or []string
-type StringMapOrSlice struct {
-	Map   map[string]string
-	Slice []string
-}
-
-// UnmarshalYAML implements custom unmarshaling for StringMapOrSlice
-func (s *StringMapOrSlice) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var m map[string]string
-	if err := unmarshal(&m); err == nil {
-		s.Map = m
-		return nil
-	}
-
-	var sl []string
-	if err := unmarshal(&sl); err == nil {
-		s.Slice = sl
-		return nil
-	}
-
-	return nil
-}
-
-// MarshalYAML implements custom marshaling for StringMapOrSlice
-func (s StringMapOrSlice) MarshalYAML() (interface{}, error) {
-	if s.Map != nil {
-		return s.Map, nil
-	}
-	return s.Slice, nil
-}
-
-// NumberOrString can be either a number or a string
-type NumberOrString struct {
-	Number *float64
-	String *string
-}
-
-// UnmarshalYAML implements custom unmarshaling for NumberOrString
-func (n *NumberOrString) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var num float64
-	if err := unmarshal(&num); err == nil {
-		n.Number = &num
-		return nil
-	}
-
-	var s string
-	if err := unmarshal(&s); err == nil {
-		n.String = &s
-		return nil
-	}
-
-	return nil
-}
-
-// MarshalYAML implements custom marshaling for NumberOrString
-func (n NumberOrString) MarshalYAML() (interface{}, error) {
-	if n.Number != nil {
-		return *n.Number, nil
-	}
-	return n.String, nil
-}
-
 // UlimitValue can be either a number or a "soft:hard" string
 type UlimitValue = NumberOrString
-
-// ParallelStep represents a step that runs steps in parallel
-type ParallelStep struct {
-	OptionalStepBase `yaml:",inline" json:",inline"`
-	// Parallel are the steps to run in parallel
-	Parallel []Step `yaml:"parallel" json:"parallel" jsonschema:"required"`
-}
-
-// SequenceStep represents a step that runs steps in sequence
-type SequenceStep struct {
-	OptionalStepBase `yaml:",inline" json:",inline"`
-	// Sequence are the steps to run in sequence
-	Sequence []Step `yaml:"sequence" json:"sequence" jsonschema:"required"`
-}
 
 // StepOrSteps can be a single step or an array of steps
 type StepOrSteps struct {
@@ -179,17 +72,6 @@ func (s StepOrSteps) MarshalYAML() (interface{}, error) {
 		return s.Single, nil
 	}
 	return s.Multiple, nil
-}
-
-// SwitchStep represents a step that switches on a value
-type SwitchStep struct {
-	OptionalStepBase `yaml:",inline" json:",inline"`
-	// Switch is the expression to switch on
-	Switch string `yaml:"switch" json:"switch" jsonschema:"required"`
-	// Case are the case branches
-	Case map[string]StepOrSteps `yaml:"case" json:"case" jsonschema:"required"`
-	// Default is the default branch
-	Default *StepOrSteps `yaml:"default,omitempty" json:"default,omitempty"`
 }
 
 // Step represents any step type (discriminated union)
