+++ pkg/schema/watch.go
@@ -82,64 +82,3 @@ func (w Watch) MarshalYAML() (interface{}, error) {
 	}
 	return nil, nil
 }
-
-// IsEnabled returns true if watch mode is enabled
-func (w *Watch) IsEnabled() bool {
-	if w == nil {
-		return false
-	}
-	if w.Bool != nil {
-		return *w.Bool
-	}
-	// Any other variant means watch is enabled
-	return w.String != nil || w.Strings != nil || w.Config != nil
-}
-
-// GetFiles returns the glob patterns to watch
-func (w *Watch) GetFiles() []string {
-	if w == nil {
-		return nil
-	}
-	if w.String != nil {
-		return []string{*w.String}
-	}
-	if w.Strings != nil {
-		return w.Strings
-	}
-	if w.Config != nil {
-		return w.Config.Files
-	}
-	return nil
-}
-
-// GetMode returns the watch mode (defaults to rebuild-reload)
-func (w *Watch) GetMode() WatchMode {
-	if w != nil && w.Config != nil && w.Config.Mode != "" {
-		return w.Config.Mode
-	}
-	return WatchModeRebuildReload
-}
-
-// ShouldUseGitIgnore returns whether to respect .gitignore (default: true)
-func (w *Watch) ShouldUseGitIgnore() bool {
-	if w != nil && w.Config != nil && w.Config.UseGitIgnore != nil {
-		return *w.Config.UseGitIgnore
-	}
-	return true
-}
-
-// ShouldUseDockerIgnore returns whether to respect .dockerignore (default: true)
-func (w *Watch) ShouldUseDockerIgnore() bool {
-	if w != nil && w.Config != nil && w.Config.UseDockerIgnore != nil {
-		return *w.Config.UseDockerIgnore
-	}
-	return true
-}
-
-// GetIgnorePatterns returns additional ignore patterns
-func (w *Watch) GetIgnorePatterns() []string {
-	if w != nil && w.Config != nil {
-		return w.Config.Ignore
-	}
-	return nil
-}
